+++ delta/layers/__init__.py
@@ -20,8 +20,10 @@ from delta.layers.recurrent import RnnAttentionEncoder
 from delta.layers.attention import MatchAttention
 from delta.layers.recurrent import RnnEncoder
 from delta.layers.recurrent import RnnDecoder
-from delta.layers.transformer import PositionEmbedding
-from delta.layers.transformer import TransformerEncoder
-from delta.layers.transformer import TransformerDecoder
+from delta.layers.transformer_sub import MultiHeadAttention
+from delta.layers.transformer_sub import PositionEmbedding
+from delta.layers.transformer_sub import PositionwiseFeedForward
+from delta.layers.transformer_old import TransformerEncoder
+from delta.layers.transformer_old import TransformerDecoder
 
 from delta.layers.common_layers import *
